refactor: update migration components and streamline imports
- Introduced `MigrateInfoScreen` and `MigrateSuccess` components to enhance the migration flow.
- Updated `VaultKeygenEnding` and `MigrateIntro` to import these new components from the core path, improving consistency and maintainability.
- Cleaned up obsolete import paths to align with the new structure.
