+++ clients/desktop/src/vault/keygen/shared/VaultKeygenEnding.tsx
@@ -1,9 +1,9 @@
+import { MigrateSuccess } from '@core/ui/mpc/keygen/migrate/MigrateSuccess'
 import { useCurrentKeygenType } from '@core/ui/mpc/keygen/state/currentKeygenType'
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { StepTransition } from '@lib/ui/base/StepTransition'
 import { OnBackProp } from '@lib/ui/props'
 
-import { MigrateSuccess } from '../../migrate/MigrateSuccess'
 import { VaultKeygenBackupFlow } from './VaultKeygenBackupFlow'
 
 export const VaultKeygenEnding = ({ onBack }: OnBackProp) => {
