+++ clients/desktop/src/vault/migrate/MigrateIntro.tsx
@@ -1,11 +1,10 @@
+import { MigrateInfoScreen } from '@core/ui/mpc/keygen/migrate/MigrateInfoScreen'
 import { StepTransition } from '@lib/ui/base/StepTransition'
 import { Button } from '@lib/ui/buttons/Button'
 import { ChevronRightIcon } from '@lib/ui/icons/ChevronRightIcon'
 import { HStack } from '@lib/ui/layout/Stack'
 import { OnFinishProp } from '@lib/ui/props'
 import { Trans, useTranslation } from 'react-i18next'
-
-import { MigrateInfoScreen } from './MigrateInfoScreen'
 
 export const MigrateIntro = ({ onFinish }: OnFinishProp) => {
   const { t } = useTranslation()
