+++ core/ui/mpc/keygen/migrate/MigrateInfoScreen.tsx
