+++ core/ui/mpc/keygen/migrate/MigrateSuccess.tsx
